requirements.txt
streamlit>=1.35.0
pymupdf>=1.24.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
anthropic>=0.28.0
python-dotenv>=1.0.0
